\name{computer}
\Rdversion{1.1}
\alias{computer}
\docType{data}
\title{
  Repair time of computers
}
\description{
  Data from a study of the repair time for different brands of computers
  in different configurations.
}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{brand}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{type}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{time}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(computer)
}
\keyword{datasets}
