\name{deink}
\Rdversion{1.1}
\alias{deink}
\docType{data}
\title{
  De-inking of newsprint
}
\description{
  De-inking of newsprint
}
\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{alkali}}{a numeric vector}
    \item{\code{hardness}}{a numeric vector}
    \item{\code{bright}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(deink)
}
\keyword{datasets}
