\name{deink2}
\Rdversion{1.1}
\alias{deink2}
\docType{data}
\title{
  De-inking of newsprint
}
\description{
  De-inking of newsprint
}
\format{
  A data frame with 15 observations on the following 3 variables.
  \describe{
    \item{\code{formula}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{newspaper}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E}}
    \item{\code{bright}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(deink2)
}
\keyword{datasets}
