\name{diagnostic}
\Rdversion{1.1}
\alias{diagnostic}
\docType{data}
\title{
 Consistency of diagnostic kits
}
\description{
 Consistency of diagnostic kit
}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{kit1}}{a numeric vector}
    \item{\code{kit2}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(diagnostic)
}
\keyword{datasets}
