\name{drums}
\Rdversion{1.1}
\alias{drums}
\docType{data}
\title{
  Weights of drums before and after filling
}
\description{
  Weights of drums before and after being filled with a chemical product.
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{number}}{a numeric vector}
    \item{\code{empty}}{a numeric vector}
    \item{\code{full}}{a numeric vector}
  }
}
\details{
  These data were collected to study a filling process that filled drums
  of material based on an assumed weight of the empty drum.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(drums)
}
\keyword{datasets}
