\name{epoxy}
\Rdversion{1.1}
\alias{epoxy}
\docType{data}
\title{
  Effect of epoxy level on appearance
}
\description{
  Effect of epoxy level on appearance
}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{epoxy}}{level of epoxy in formulation}
    \item{\code{lw}}{longwave measure of appearance}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(epoxy)
}
\keyword{datasets}
