\name{fbuild}
\Rdversion{1.1}
\alias{fbuild}
\docType{data}
\title{
  Appearance for film builds
}
\description{
  Appearance for film builds
}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{build}}{a numeric vector}
    \item{\code{gloss}}{a numeric vector}
  }
}
\details{
  Film build is the thickness of a coating, such as an automotive paint.
  Gloss is an appearance measure with higher gloss generally being
  preferred.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(fbuild)
}
\keyword{datasets}
