\name{fill}
\Rdversion{1.1}
\alias{fill}
\docType{data}
\title{
  Fill amount in tamped cylinders
}
\description{
  Fill amount in tamped cylinders
}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{distance}}{a numeric vector}
    \item{\code{tamps}}{a numeric vector}
    \item{\code{fill}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(fill)
}
\keyword{datasets}
