\name{fillweight}
\Rdversion{1.1}
\alias{fillweight}
\docType{data}
\title{
  Fill weight by batch
}
\description{
  Fill weight by batch
}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{run}}{a factor with levels \code{A} \code{B} \code{C} \code{D}}
    \item{\code{tube}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E}}
    \item{\code{weight}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(fillweight)
}
\keyword{datasets}
