\name{gloss}
\Rdversion{1.1}
\alias{gloss}
\docType{data}
\title{
  Gloss of paint on cars
}
\description{
  Gloss of paint on cars
}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{color}}{a factor with levels \code{Black} \code{Green} \code{Silver} \code{White}}
    \item{\code{gloss}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(gloss)
}
\keyword{datasets}
