\name{lwsw}
\Rdversion{1.1}
\alias{lwsw}
\docType{data}
\title{
  Appearance measures of automotive paints
}
\description{
  Appearance measures of automotive paints
}
\format{
  A data frame with 13 observations on the following 2 variables.
  \describe{
    \item{\code{lw}}{a numeric vector}
    \item{\code{sw}}{a numeric vector}
  }
}
\details{
  Longwave (LW) and shortwave (SW) are two surface quality measurements.
  Both are important in the development of coating formulations.  This
  data was gathered to examine the relationship between the two measures.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(lwsw)
}
\keyword{datasets}
