\name{moisture}
\Rdversion{1.1}
\alias{moisture}
\docType{data}
\title{
  Moisture content of a silica product
}
\description{
  Moisture content of a silica product
}
\format{
  A data frame with 8 observations on the following 5 variables.
  \describe{
    \item{\code{temp}}{a numeric vector}
    \item{\code{speed}}{a factor with levels \code{A} \code{B}}
    \item{\code{solids}}{a numeric vector}
    \item{\code{ph}}{a numeric vector}
    \item{\code{moisture}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(moisture)
}
\keyword{datasets}
