\name{mw}
\Rdversion{1.1}
\alias{mw}
\docType{data}
\title{
  Appearance measure of paint
}
\description{
  Appearance measure of paint
}
\format{
  A data frame with 32 observations on the following 6 variables.
  \describe{
    \item{\code{a}}{a factor with levels \code{H} \code{L}}
    \item{\code{b}}{a factor with levels \code{H} \code{L}}
    \item{\code{c}}{a factor with levels \code{H} \code{L}}
    \item{\code{d}}{a factor with levels \code{H} \code{L}}
    \item{\code{e}}{a factor with levels \code{H} \code{L}}
    \item{\code{mw}}{a numeric vector}
  }
}
\details{
  These data are from a designed experiment to study the impact of 5
  process factors on the molecular weight of a paint.  The Mw is the
  response of interest as it in turn impacts quality characteristics of
  the paint. This study was run to better understand the influence of
  processing variables on the resulting Mw of the formulations.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(mw)
}
\keyword{datasets}
