\name{odor}
\Rdversion{1.1}
\alias{odor}
\docType{data}
\title{
  Odor, yellowing and hardness of optical coating
}
\description{
  Odor, yellowing and hardness of optical coating
}
\format{
  A data frame with 35 observations on the following 3 variables.
  \describe{
    \item{\code{odor}}{a numeric vector}
    \item{\code{yellowing}}{a numeric vector}
    \item{\code{hardness}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(odor)
}
\keyword{datasets}
