\name{oven}
\Rdversion{1.1}
\alias{oven}
\docType{data}
\title{
  Drying of silica
}
\description{
  Drying of silica
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{brand}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{type}}{a factor with levels \code{A} \code{B}}
    \item{\code{moisture}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(oven)
}
\keyword{datasets}
