\name{ph}
\Rdversion{1.1}
\alias{ph}
\docType{data}
\title{
  pH measurements in a chemical process
}
\description{
  pH measurements in a chemical process
}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{batch}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{ph}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(ph)
}
\keyword{datasets}
