\name{phmeas}
\Rdversion{1.1}
\alias{phmeas}
\docType{data}
\title{
  Comparison of instruments to measure pH
}
\description{
  Comparison of instruments to measure pH
}
\format{
  A data frame with 11 observations on the following 2 variables.
  \describe{
    \item{\code{phold}}{a numeric vector}
    \item{\code{phnew}}{a numeric vector}
  }
}
\details{
  These data were collected to compare an old instrument to a new
  instrument.  Eleven samples were tested with both instruments.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(phmeas)
}
\keyword{datasets}
