\name{protein}
\Rdversion{1.1}
\alias{protein}
\docType{data}
\title{
  Assay of protein in blood
}
\description{
  Assay of protein in blood
}
\format{
  A data frame with 54 observations on the following 5 variables.
  \describe{
    \item{\code{gender}}{a factor with levels \code{F} and \code{M}}
    \item{\code{age}}{a numeric vector}
    \item{\code{duration}}{a numeric vector}
    \item{\code{levela}}{a numeric vector}
    \item{\code{protein}}{a numeric vector}
  }
}
\details{
  The response of interest in this data set is the protein level from a
  new diagnostic test.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(protein)
}
\keyword{datasets}
