\name{purity}
\Rdversion{1.1}
\alias{purity}
\docType{data}
\title{
  Purity of product from a filtration process
}
\description{
  Purity of product from a filtration process
}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{
    \item{\code{tech}}{indicator factor for technician}
    \item{\code{fac1}}{a factor with 3 levels}
    \item{\code{fac2}}{a factor with 2 levels}
    \item{\code{purity}}{a response factor}
  }
}
\details{
  This experiment was run by two technicians to study the impact of two
  process factors on the purity of the product after filtration.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(purity)
}
\keyword{datasets}
