\name{railcar}
\Rdversion{1.1}
\alias{railcar}
\alias{railcar2}
\docType{data}
\title{
  Rail car hold times
}
\description{
  Rail car hold times
}
\format{
  A data frame with 53 observations on the following variable.
  \describe{
    \item{\code{days}}{a numeric vector}
  }
}
\details{
  A company ships product to customers in rail cars.  The rail cars are
  delivered by the railroad to a rail siding at the customer's facility.
  This data records the time that a railcar is held at a customer
  facility.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(railcar)
str(railcar2)
}
\keyword{datasets}
