\name{railcar3}
\Rdversion{1.1}
\alias{railcar3}
\docType{data}
\title{
  Moisture level versus type of rail car
}
\description{
  Moisture level versus type of rail car
}
\format{
  A data frame with 17 observations on the following 2 variables.
  \describe{
    \item{\code{type}}{a factor with levels \code{A} \code{B}}
    \item{\code{moisture}}{a numeric vector}
  }
}
\details{
  Rail cars are used to ship bulk product to customers.  A company has
  two primary styles of cars that it uses for a product that has
  specific moisture specifications.  A customer was seeing variability
  in the moisture levels of the product when it reached their site.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(railcar3)
}
\keyword{datasets}
