\name{ratings}
\Rdversion{1.1}
\alias{ratings}
\docType{data}
\title{
  Ratings of raw materials
}
\description{
  Ratings of raw materials
}
\format{
  A data frame with 26 observations on the following variable.
  \describe{
    \item{\code{rating}}{a numeric vector}
  }
}
\details{
  A company has a new raw material that they wish to quickly evaluate.
  To do so they rate how well their process is running at key steps when
  using the new material.  The rating is done on a scale of 1 - 10 at
  each key step and then averaged.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(ratings)
}
\keyword{datasets}
