\name{sales}
\Rdversion{1.1}
\alias{sales}
\docType{data}
\title{
  Sales versus capital expenditure
}
\description{
  Sales versus capital expenditure over 48 months
}
\format{
  A data frame with 48 observations (ordered by month) on the following
  3 variables.
  \describe{
    \item{\code{capital}}{a numeric vector}
    \item{\code{pindex}}{a numeric vector}
    \item{\code{sales}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(sales)
}
\keyword{datasets}
