\name{soap}
\Rdversion{1.1}
\alias{soap}
\docType{data}
\title{
  Soap sales by packaging type
}
\description{
  Soap sales by packaging type over a 5 week study
}
\format{
  A data frame with 5 observations on the following 3 variables.
  \describe{
    \item{\code{box}}{number of box packages sold}
    \item{\code{foil}}{number of foil packages sold}
    \item{\code{shrink}}{number of shrink wrapped packages sold}
  }
}
\details{
  An in-store study was run to test consumer preference of packaging for
  a soap.  The study was run for 5 weeks with the number of packages
  sold each week recorded.  Note, after the third week there was a
  supply disruption so there were no boxed soaps available for sale.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(soap)
}
\keyword{datasets}
