\name{stab}
\Rdversion{1.1}
\alias{stab}
\docType{data}
\title{
  Stability of a chemical product
}
\description{
  Stability of a chemical product
}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{
    \item{\code{temp}}{a numeric vector}
    \item{\code{humidity}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(stab)
}
\keyword{datasets}
