\name{stretch}
\Rdversion{1.1}
\alias{stretch}
\docType{data}
\title{
  Stretch of hot pizza cheese
}
\description{
  Stretch of hot pizza cheese
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{temp}}{a numeric vector}
    \item{\code{cheese}}{a numeric vector}
    \item{\code{stretch}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(stretch)
}
\keyword{datasets}
