\name{tablets}
\Rdversion{1.1}
\alias{tablets}
\docType{data}
\title{
  Lifetime of chlorine tablets
}
\description{
  Lifetime of chlorine tablets
}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{batch}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{time}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(tablets)
}
\keyword{datasets}
