\name{temprate}
\Rdversion{1.1}
\alias{temprate}
\docType{data}
\title{
  Effect of water bath on moisture content
}
\description{
  Effect of water bath on moisture content
}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{temp}}{a numeric vector}
    \item{\code{rate}}{a factor with levels \code{FAST} \code{MED} \code{SLOW}}
    \item{\code{response}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(temprate)
}
\keyword{datasets}
