\name{tpaste}
\Rdversion{1.1}
\alias{tpaste}
\docType{data}
\title{
  Turbidity of a toothpaste formulation
}
\description{
  Turbidity of a toothpaste formulation
}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{nacl}}{a numeric vector}
    \item{\code{temp}}{a numeric vector}
    \item{\code{rate}}{a numeric vector}
    \item{\code{turbidity}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(tpaste)
}
\keyword{datasets}
