\name{urine}
\Rdversion{1.1}
\alias{urine}
\docType{data}
\title{
  Mercury level in employee urine samples
}
\description{
  Mercury level in employee urine samples
}
\format{
  A data frame with 12 observations on the following 5 variables.
  \describe{
    \item{\code{month.s10.}}{a factor with levels \code{April} \code{August} \code{Decmber} \code{February} \code{January} \code{July} \code{June} \code{March} \code{May} \code{November} \code{October} \code{September}}
    \item{\code{person1}}{a numeric vector}
    \item{\code{person2}}{a numeric vector}
    \item{\code{person3}}{a numeric vector}
    \item{\code{person4}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(urine)
}
\keyword{datasets}
