\name{uvoven}
\Rdversion{1.1}
\alias{uvoven}
\docType{data}
\title{
  UV absorbance for lens cured in different ovens
}
\description{
  UV absorbance for lens cured in different ovens
}
\format{
  A data frame with 60 observations on the following 2 variables.
  \describe{
    \item{\code{oven}}{a factor with levels \code{A} \code{B}}
    \item{\code{uv}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(uvoven)
}
\keyword{datasets}
