\name{vitamin}
\Rdversion{1.1}
\alias{vitamin}
\docType{data}
\title{
  Calcium levels before and after vitamin supplement
}
\description{
  Calcium levels before and after vitamin supplement
}
\format{
  A data frame with 49 observations on the following 3 variables.
  \describe{
    \item{\code{treatment}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{before}}{a numeric vector}
    \item{\code{after}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(vitamin)
}
\keyword{datasets}
