\name{wash}
\Rdversion{1.1}
\alias{wash}
\docType{data}
\title{
  Appearance of washed and unwashed panels
}
\description{
  Appearance of washed and unwashed panels
}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{type}}{a factor with levels \code{NW} \code{W}}
    \item{\code{time1}}{a numeric vector}
    \item{\code{time2}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(wash)
}
\keyword{datasets}
