\name{water}
\Rdversion{1.1}
\alias{water}
\docType{data}
\title{
  Bacteria concentrations in water samples
}
\description{
  Bacteria concentrations in water samples
}
\format{
  A data frame with 50 observations on the following variable.
  \describe{
    \item{\code{bacteria}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(water)
}
\keyword{datasets}
