\name{webtraff}
\Rdversion{1.1}
\alias{webtraff}
\docType{data}
\title{
  Web site traffic during a marketing campaign
}
\description{
  Web site traffic during a marketing campaign
}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{weeks}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E}}
    \item{\code{traffic}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(webtraff)
}
\keyword{datasets}
