\name{yield}
\Rdversion{1.1}
\alias{yield}
\docType{data}
\title{
  Yield of a chemical process
}
\description{
  Yield of a chemical process
}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{temp}}{a numeric vector}
    \item{\code{ph}}{a numeric vector}
    \item{\code{yield}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(yield)
}
\keyword{datasets}
