\name{dry}
\Rdversion{1.1}
\alias{dry}
\docType{data}
\title{
  Moisture content after drying
}
\description{
  Moisture content after drying
}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{rate}}{a factor with levels \code{SLOW} and  \code{FAST}}
    \item{\code{type}}{type of dryer - a factor with levels
      \code{BATCH} and \code{TUNNEL}}
    \item{\code{response}}{Moisture content of final product (\%)}
  }
}
\examples{
str(dry)
}
\keyword{datasets}
