\name{fish}
\Rdversion{1.1}
\alias{fish}
\alias{fish2}
\docType{data}
\title{
  Toxin concentrations in fish by site
}
\description{
  Toxin concentrations in fish by site
}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{site}}{a factor with levels \code{A} \code{B} \code{C} \code{D}}
    \item{\code{toxin}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(fish)
str(fish2)
}
\keyword{datasets}
