\name{pigment}
\Rdversion{1.1}
\alias{pigment}
\docType{data}
\title{
  Yellowing of paint for different pigments
}
\description{
  Yellowing of paint for different pigments
}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{batch}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{pigment}}{a factor with levels \code{A} \code{B} \code{C}}
    \item{\code{deltab}}{a numeric vector}
  }
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(pigment)
}
\keyword{datasets}
