\name{stretch}
\Rdversion{1.1}
\alias{stretch}
\docType{data}
\title{
  Stretch of hot pizza cheese
}
\description{
  Stretch of hot pizza cheese
}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{temp}}{a numeric vector}
    \item{\code{cheese}}{a numeric vector}
    \item{\code{stretch}}{a numeric vector}
  }
}
\details{
  A critical quality measure of pizza cheese is how well it stretches
  when it is hot.  This study was to evaluate the amount of cheese and
  the temperature at which the pizza was cooked on the stretch of the
  hot cheese.  Measuring the stretch of pizza cheese is not a precise
  science, thus, five measurements were taken on each pizza.  One pizza
  was used for each temperature/cheese combination.
}
\references{
  Peter R. Nelson, Marie Coffin and Karen A. F. Copeland (2003),
  \emph{Introductory Statistics for Engineering Experimentation},
  Elsevier. (Appendix A)
}
\examples{
str(stretch)
}
\keyword{datasets}
